Create LavfiMetadataFilterParser and move it all LavfiMtdProgram Frames and Events Extractors #33
