+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/AudioFilterAPhasemeter.java
@@ -17,25 +17,15 @@
 package tv.hd3g.fflauncher.filtering;
 
 import java.time.Duration;
-import java.util.List;
-import java.util.Optional;
 
 import lombok.Data;
-import tv.hd3g.fflauncher.filtering.AudioFilterAPhasemeter.LavfiMtdAPhaseMeter;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEvents;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEventsExtractor;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramFrames;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramFramesExtractor;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiRawMtdFrame;
 
 /**
  * https://www.ffmpeg.org/ffmpeg-filters.html#phasing-detection
  */
 @Data
-public class AudioFilterAPhasemeter implements AudioFilterSupplier,
-									LavfiMtdProgramFramesExtractor<LavfiMtdAPhaseMeter>,
-									LavfiMtdProgramEventsExtractor {
-	private static final String APHASEMETER = "aphasemeter";
+public class AudioFilterAPhasemeter implements AudioFilterSupplier {
+	public static final String APHASEMETER = "aphasemeter";
 
 	private boolean phasing;
 	private boolean video;
@@ -59,31 +49,6 @@ public class AudioFilterAPhasemeter implements AudioFilterSupplier,
 			f.addArgument("video", "0");
 		}
 		return f;
-	}
-
-	/**
-	 * frame:1022 pts:981168 pts_time:20.441
-	 * lavfi.aphasemeter.phase=1.000000
-	 * lavfi.aphasemeter.mono_start=18.461
-	 * frame:1299 pts:1247088 pts_time:25.981
-	 * lavfi.aphasemeter.phase=0.992454
-	 * lavfi.aphasemeter.mono_end=25.981
-	 * lavfi.aphasemeter.mono_duration=2.94
-	 */
-	@Override
-	public LavfiMtdProgramFrames<LavfiMtdAPhaseMeter> getMetadatas(final List<? extends LavfiRawMtdFrame> extractedRawMtdFrames) {
-		return new LavfiMtdProgramFrames<>(extractedRawMtdFrames, APHASEMETER,
-				rawFrames -> Optional.ofNullable(rawFrames.get("phase"))
-						.map(Float::parseFloat)
-						.map(LavfiMtdAPhaseMeter::new));
-	}
-
-	public record LavfiMtdAPhaseMeter(float phase) {
-	}
-
-	@Override
-	public LavfiMtdProgramEvents getEvents(final List<? extends LavfiRawMtdFrame> extractedRawMtdFrames) {
-		return new LavfiMtdProgramEvents(extractedRawMtdFrames, APHASEMETER, "mono");
 	}
 
 }
