+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/AudioFilterSilencedetect.java
@@ -17,21 +17,11 @@
 package tv.hd3g.fflauncher.filtering;
 
 import java.time.Duration;
-import java.util.List;
 
 import lombok.Data;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEvents;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEventsExtractor;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiRawMtdFrame;
 
-/**
- * https://www.ffmpeg.org/ffmpeg-filters.html#silencedetect
- * lavfi.silence_start/lavfi.silence_start.X
- * lavfi.silence_duration/lavfi.silence_duration.X,
- * lavfi.silence_end/lavfi.silence_end.X
- */
 @Data
-public class AudioFilterSilencedetect implements AudioFilterSupplier, LavfiMtdProgramEventsExtractor {
+public class AudioFilterSilencedetect implements AudioFilterSupplier {
 
 	private boolean mono;
 	private Duration duration;
@@ -51,21 +41,6 @@ public class AudioFilterSilencedetect implements AudioFilterSupplier, LavfiMtdPr
 		f.addOptionalNonNegativeArgument("noise", noiseDb);
 		f.addOptionalNonNegativeArgument("noise", noiseRatio);
 		return f;
-	}
-
-	/**
-	 * frame:3306 pts:3173808 pts_time:66.121
-	 * lavfi.silence_start.1=65.1366
-	 * lavfi.silence_start.2=65.1366
-	 * frame:3332 pts:3198768 pts_time:66.641
-	 * lavfi.silence_end.1=66.6474
-	 * lavfi.silence_duration.1=1.51079
-	 * lavfi.silence_end.2=66.6474
-	 * lavfi.silence_duration.2=1.51079 *
-	 */
-	@Override
-	public LavfiMtdProgramEvents getEvents(final List<? extends LavfiRawMtdFrame> extractedRawMtdFrames) {
-		return new LavfiMtdProgramEvents(extractedRawMtdFrames, "silence");
 	}
 
 }
