+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/VideoFilterBlackdetect.java
@@ -17,18 +17,14 @@
 package tv.hd3g.fflauncher.filtering;
 
 import java.time.Duration;
-import java.util.List;
 
 import lombok.Data;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEvents;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEventsExtractor;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiRawMtdFrame;
 
 /**
  * https://www.ffmpeg.org/ffmpeg-filters.html#blackdetect
  */
 @Data
-public class VideoFilterBlackdetect implements VideoFilterSupplier, LavfiMtdProgramEventsExtractor {
+public class VideoFilterBlackdetect implements VideoFilterSupplier {
 
 	private Duration blackMinDuration;
 	private float pictureBlackRatioTh;
@@ -46,21 +42,6 @@ public class VideoFilterBlackdetect implements VideoFilterSupplier, LavfiMtdProg
 		f.addOptionalNonNegativeArgument("picture_black_ratio_th", pictureBlackRatioTh);
 		f.addOptionalNonNegativeArgument("pixel_black_th", pixelBlackTh);
 		return f;
-	}
-
-	/**
-	 * frame:0 pts:7 pts_time:0.007
-	 * lavfi.black_start=0.007
-	 * frame:2 pts:90 pts_time:0.09
-	 * lavfi.black_end=0.09
-	 * frame:108 pts:4507 pts_time:4.507
-	 * lavfi.black_start=4.507
-	 * frame:116 pts:4841 pts_time:4.841
-	 * lavfi.black_end=4.841
-	 */
-	@Override
-	public LavfiMtdProgramEvents getEvents(final List<? extends LavfiRawMtdFrame> extractedRawMtdFrames) {
-		return new LavfiMtdProgramEvents(extractedRawMtdFrames, "black");
 	}
 
 }
