+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/VideoFilterFreezedetect.java
@@ -17,22 +17,14 @@
 package tv.hd3g.fflauncher.filtering;
 
 import java.time.Duration;
-import java.util.List;
 
 import lombok.Data;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEvents;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdProgramEventsExtractor;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiRawMtdFrame;
 
 /**
  * https://www.ffmpeg.org/ffmpeg-filters.html#freezedetect
- * lavfi.freezedetect.freeze_start
- * lavfi.freezedetect.freeze_duration
- * lavfi.freezedetect.freeze_end
- * No thread safe
  */
 @Data
-public class VideoFilterFreezedetect implements VideoFilterSupplier, LavfiMtdProgramEventsExtractor {
+public class VideoFilterFreezedetect implements VideoFilterSupplier {
 
 	private float noiseToleranceRatio;
 	private int noiseToleranceDb;
@@ -50,18 +42,6 @@ public class VideoFilterFreezedetect implements VideoFilterSupplier, LavfiMtdPro
 		f.addOptionalArgument("noise", noiseToleranceDb > Integer.MIN_VALUE, "-" + Math.abs(noiseToleranceDb) + "dB");
 		f.addOptionalNonNegativeArgument("noise", noiseToleranceRatio);
 		return f;
-	}
-
-	/**
-	 * frame:66 pts:2757 pts_time:2.757
-	 * lavfi.freezedetect.freeze_start=0.757
-	 * frame:97 pts:4049 pts_time:4.049
-	 * lavfi.freezedetect.freeze_duration=3.292
-	 * lavfi.freezedetect.freeze_end=4.049
-	 */
-	@Override
-	public LavfiMtdProgramEvents getEvents(final List<? extends LavfiRawMtdFrame> extractedRawMtdFrames) {
-		return new LavfiMtdProgramEvents(extractedRawMtdFrames, "freezedetect", "freeze");
 	}
 
 }
