+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/lavfimtd/LavfiMtdCropdetect.java
@@ -11,27 +11,10 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU Lesser General Public License for more details.
  *
- * Copyright (C) hdsdi3g for hd3g.tv 2022
+ * Copyright (C) hdsdi3g for hd3g.tv 2023
  *
  */
 package tv.hd3g.fflauncher.filtering.lavfimtd;
 
-import java.util.Map;
-
-public interface LavfiRawMtdFrame {
-
-	String DEFAULT_KEY = "default";
-
-	int getFrame();
-
-	long getPts();
-
-	float getPtsTime();
-
-	Map<String, Map<String, String>> getValuesByFilterKeysByFilterName();
-
-	default LavfiMtdPosition getLavfiMtdPosition() {
-		return new LavfiMtdPosition(getFrame(), getPts(), getPtsTime());
-	}
-
+public record LavfiMtdCropdetect(int x1, int x2, int y1, int y2, int w, int h, int x, int y) {
 }
