+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/lavfimtd/LavfiMtdIdetRepeatedFrame.java
@@ -11,15 +11,13 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU Lesser General Public License for more details.
  *
- * Copyright (C) hdsdi3g for hd3g.tv 2022
+ * Copyright (C) hdsdi3g for hd3g.tv 2023
  *
  */
 package tv.hd3g.fflauncher.filtering.lavfimtd;
 
-import java.util.List;
-
-public interface LavfiMtdProgramEventsExtractor {
-
-	LavfiMtdProgramEvents getEvents(List<? extends LavfiRawMtdFrame> extractedRawMtdFrames);
-
+public record LavfiMtdIdetRepeatedFrame(LavfiMtdIdetRepeatedFrameType currentFrame,
+										int neither,
+										int top,
+										int bottom) {
 }
