+++ fflauncher/src/main/java/tv/hd3g/fflauncher/filtering/lavfimtd/LavfiMtdPosition.java
@@ -11,12 +11,12 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU Lesser General Public License for more details.
  *
- * Copyright (C) hdsdi3g for hd3g.tv 2022
+ * Copyright (C) hdsdi3g for hd3g.tv 2023
  *
  */
 package tv.hd3g.fflauncher.filtering.lavfimtd;
 
-public record LavfiMtdPosition(int frame, long pts, float ptsTime) implements Comparable<LavfiMtdPosition> {
+record LavfiMtdPosition(int frame, long pts, float ptsTime) implements Comparable<LavfiMtdPosition> {
 
 	@Override
 	public int compareTo(final LavfiMtdPosition o) {
