+++ fflauncher/src/test/java/tv/hd3g/fflauncher/filtering/AudioFilterSilencedetectTest.java
@@ -17,28 +17,11 @@
 package tv.hd3g.fflauncher.filtering;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static tv.hd3g.fflauncher.filtering.lavfimtd.Utility.getFramesFromString;
-
-import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdEvent;
-
 class AudioFilterSilencedetectTest {
-
-	private static final String RAW_LINES = """
-			frame:3306 pts:3173808 pts_time:66.121
-			lavfi.silence_start.1=65.1366
-			lavfi.silence_start.2=65.1367
-			frame:3332 pts:3198768 pts_time:66.641
-			lavfi.silence_end.1=66.6474
-			lavfi.silence_duration.1=1.51079
-			lavfi.silence_end.2=66.6479
-			lavfi.silence_duration.2=1.51079 *
-			""";
 
 	AudioFilterSilencedetect f;
 
@@ -52,13 +35,4 @@ class AudioFilterSilencedetectTest {
 		assertEquals("silencedetect", f.toFilter().toString());
 	}
 
-	@Test
-	void testGetEvents() {
-		final var events = f.getEvents(getFramesFromString(RAW_LINES));
-		assertNotNull(events);
-		assertEquals(List.of(
-				new LavfiMtdEvent("silence", "1", 65.1366f, 66.6474f),
-				new LavfiMtdEvent("silence", "2", 65.1367f, 66.6479f)),
-				events.getEvents());
-	}
 }
