+++ fflauncher/src/test/java/tv/hd3g/fflauncher/filtering/VideoFilterBlurdetectTest.java
@@ -17,23 +17,11 @@
 package tv.hd3g.fflauncher.filtering;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static tv.hd3g.fflauncher.filtering.lavfimtd.Utility.getFramesFromString;
-
-import java.util.Map;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import tv.hd3g.fflauncher.filtering.VideoFilterBlurdetect.LavfiMtdBlurdetect;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdPosition;
-
 class VideoFilterBlurdetectTest {
-
-	private static final String RAW_LINES = """
-			frame:114 pts:4757 pts_time:4.757
-			lavfi.blur=5.744382
-			""";
 
 	VideoFilterBlurdetect f;
 
@@ -45,15 +33,6 @@ class VideoFilterBlurdetectTest {
 	@Test
 	void test() {
 		assertEquals("blurdetect", f.toFilter().toString());
-	}
-
-	@Test
-	void testGetMetadatas() {
-		final var frames = f.getMetadatas(getFramesFromString(RAW_LINES));
-		assertNotNull(frames);
-		assertEquals(Map.of(
-				new LavfiMtdPosition(114, 4757, 4.757f), new LavfiMtdBlurdetect(5.744382f)),
-				frames.getFrames());
 	}
 
 }
