+++ fflauncher/src/test/java/tv/hd3g/fflauncher/filtering/VideoFilterSitiTest.java
@@ -17,26 +17,11 @@
 package tv.hd3g.fflauncher.filtering;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static tv.hd3g.fflauncher.filtering.lavfimtd.Utility.getFramesFromString;
-
-import java.util.DoubleSummaryStatistics;
-import java.util.Map;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import tv.hd3g.fflauncher.filtering.VideoFilterSiti.LavfiMtdSiti;
-import tv.hd3g.fflauncher.filtering.VideoFilterSiti.LavfiMtdSitiSummary;
-import tv.hd3g.fflauncher.filtering.lavfimtd.LavfiMtdPosition;
-
 class VideoFilterSitiTest {
-
-	private static final String RAW_LINES = """
-			frame:119 pts:4966 pts_time:4.966
-			lavfi.siti.si=11.67
-			lavfi.siti.ti=5.60
-			""";
 
 	VideoFilterSiti f;
 
@@ -50,56 +35,4 @@ class VideoFilterSitiTest {
 		assertEquals("siti", f.toFilter().toString());
 	}
 
-	@Test
-	void testGetMetadatas() {
-		final var frames = f.getMetadatas(getFramesFromString(RAW_LINES));
-		assertNotNull(frames);
-		assertEquals(Map.of(
-				new LavfiMtdPosition(119, 4966, 4.966f), new LavfiMtdSiti(11.67f, 5.60f)),
-				frames.getFrames());
-	}
-
-	@Test
-	void testGetMetadatas_noSi() {
-		final var frames = f.getMetadatas(getFramesFromString("""
-				frame:119 pts:4966 pts_time:4.966
-				lavfi.siti.ti=5.60
-				"""));
-		assertNotNull(frames);
-		assertEquals(Map.of(), frames.getFrames());
-	}
-
-	@Test
-	void testGetMetadatas_noTi() {
-		final var frames = f.getMetadatas(getFramesFromString("""
-				frame:119 pts:4966 pts_time:4.966
-				lavfi.siti.si=5.60
-				"""));
-		assertNotNull(frames);
-		assertEquals(Map.of(), frames.getFrames());
-	}
-
-	@Test
-	void testComputeSitiStats() {
-		final var frames = f.getMetadatas(getFramesFromString("""
-				frame:118 pts:4965 pts_time:4.965
-				lavfi.siti.si=5
-				lavfi.siti.ti=7
-				frame:119 pts:4966 pts_time:4.966
-				lavfi.siti.si=15
-				lavfi.siti.ti=3
-				"""));
-		final var sum = VideoFilterSiti.computeSitiStats(frames);
-		assertNotNull(sum);
-
-		final var dssSi = new DoubleSummaryStatistics();
-		dssSi.accept(5);
-		dssSi.accept(15);
-
-		final var dssTi = new DoubleSummaryStatistics();
-		dssTi.accept(7);
-		dssTi.accept(3);
-
-		assertEquals(new LavfiMtdSitiSummary(dssSi, dssTi).toString(), sum.toString());
-	}
 }
